fix art price

+++ src/main/java/ru/vsu/cs/sheina/online_gallery_backend/dto/art/ArtChangeDTO.java
@@ -3,6 +3,7 @@ package ru.vsu.cs.sheina.online_gallery_backend.dto.art;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Data
@@ -21,7 +22,7 @@ public class ArtChangeDTO {
 
     Boolean isPrivate;
 
-    Double price;
+    BigInteger price;
 
     String description;
