+++ src/main/java/ru/vsu/cs/sheina/online_gallery_backend/dto/artist/ArtistFullDTO.java
@@ -3,6 +3,7 @@ package ru.vsu.cs.sheina.online_gallery_backend.dto.artist;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigInteger;
 import java.util.UUID;
 
 @Data
@@ -21,7 +22,7 @@ public class ArtistFullDTO {
 
     Integer countSoldArts;
 
-    Double salesAmount;
+    BigInteger salesAmount;
 
     Integer countSubscribers;
